clean0up, get rid of unnecessary init methods
