+++ src/apps/java/consumption/com.sap.iot.starterkit.ui/src/main/webapp/js/view/main.view.js
@@ -1,4 +1,3 @@
-// jQuery.sap.require( 'js.base.Controller' );
 sap.ui.jsview( "js.view.main", {
 
 	getControllerName: function() {
@@ -9,10 +8,10 @@ sap.ui.jsview( "js.view.main", {
 		this.setDisplayBlock( true );
 
 		var oOutboundView = sap.ui.jsview( "js.view.outbound" );
-		this.addDependent( oOutboundView );
+		// this.addDependent( oOutboundView );
 
 		var oInboundView = sap.ui.jsview( "js.view.inbound" );
-		this.addDependent( oInboundView );
+		// this.addDependent( oInboundView );
 
 		return new sap.m.Page( {
 			title: "{i18n>TITLE}",
